Avance autorizaciones y roles de usuarios

+++ E-Comerce-Discos-Vinilo/src/controller/carts.controller.js
@@ -14,7 +14,6 @@ export class CartsController {
   static getCarts = async (req, res) => {
     try {
       const carts = await cartsService.getCarts();
-      console.log(carts);
       res.json({ data: carts });
     } catch (error) {
       logger.error(error.message);
@@ -94,16 +93,27 @@ export class CartsController {
       const quantity = req.body.quantity;
       const cartId = req.params.cartId;
       const productId = req.params.productId;
-      const newCart = await cartsService.addProduct(
-        cartId,
-        productId,
-        quantity
-      );
-      res.json({
-        status: "success",
-        message: "Producto agregado al carrito",
-        data: newCart,
-      });
+      const product = await productsService.getProductById(productId);
+      if (
+        (req.user.role === "Premium" && product.owner === req.user.id) ||
+        req.user.role === "Administrador"
+      ) {
+        res.json({
+          status: "error",
+          message: "No puedes agregar este producto al carrito",
+        });
+      } else {
+        const newCart = await cartsService.addProduct(
+          cartId,
+          productId,
+          quantity
+        );
+        res.json({
+          status: "success",
+          message: "¡Producto agregado al carrito!",
+          data: newCart,
+        });
+      }
     } catch (error) {
       logger.error(error.message);
       res.status(500).json({ message: error.message });
@@ -167,12 +177,23 @@ export class CartsController {
     try {
       const cartId = req.params.cartId;
       const productId = req.params.productId;
-      const cart = await cartsService.deleteProduct(cartId, productId);
-      res.json({
-        status: "success",
-        message: "Producto eliminado del carrito",
-        data: cart,
-      });
+      const product = await productsService.getProductById(productId);
+      if (
+        (req.user.role !== "Premium" && product.owner === req.user.id) ||
+        req.user.role !== "Administrador"
+      ) {
+        const cart = await cartsService.deleteProduct(cartId, productId);
+        res.json({
+          status: "success",
+          message: "Producto eliminado del carrito",
+          data: cart,
+        });
+      } else {
+        res.json({
+          status: "error",
+          message: "No tienes permiso para eliminar el producto",
+        });
+      }
     } catch (error) {
       logger.error(error.message);
       res.status(500).json({ message: error.message });
